requirements.txt
flask
spotipy
google-generativeai
python-dotenv
